Add snapshot options - ability to overwrite existing options and specify new. Move code around to reduce duplciation of code.
