+++ packages/drivers/odsp-socket-storage/src/odspDocumentService.ts
@@ -3,7 +3,7 @@
  * Licensed under the MIT License.
  */
 
-import { ITelemetryBaseLogger, ITelemetryLogger } from "@fluidframework/common-definitions";
+import { ITelemetryLogger } from "@fluidframework/common-definitions";
 import {
     ChildLogger,
     PerformanceEvent,
@@ -17,14 +17,11 @@ import {
     IDocumentService,
     IResolvedUrl,
     IDocumentStorageService,
-    IDocumentServiceFactory,
 } from "@fluidframework/driver-definitions";
 import {
     IClient,
     IErrorTrackingService,
-    ISummaryTree,
 } from "@fluidframework/protocol-definitions";
-import { ensureFluidResolvedUrl } from "@fluidframework/driver-utils";
 import { IOdspResolvedUrl, ISocketStorageDiscovery } from "./contracts";
 import { createNewFluidFile } from "./createFile";
 import { debug } from "./debug";
@@ -34,7 +31,7 @@ import { OdspDeltaStorageService } from "./odspDeltaStorageService";
 import { OdspDocumentDeltaConnection } from "./odspDocumentDeltaConnection";
 import { OdspDocumentStorageManager } from "./odspDocumentStorageManager";
 import { OdspDocumentStorageService } from "./odspDocumentStorageService";
-import { getWithRetryForTokenRefresh, isLocalStorageAvailable, INewFileInfo } from "./odspUtils";
+import { getWithRetryForTokenRefresh, isLocalStorageAvailable } from "./odspUtils";
 import { fetchJoinSession } from "./vroom";
 import { isOdcOrigin } from "./odspUrlHelper";
 
@@ -63,18 +60,18 @@ export class OdspDocumentService implements IDocumentService {
         resolvedUrl: IResolvedUrl,
         getStorageToken: (siteUrl: string, refresh: boolean) => Promise<string | null>,
         getWebsocketToken: (refresh) => Promise<string | null>,
-        logger: ITelemetryBaseLogger | undefined,
+        logger: ITelemetryLogger,
         storageFetchWrapper: IFetchWrapper,
         deltasFetchWrapper: IFetchWrapper,
         socketIOClientP: Promise<SocketIOClientStatic>,
         cache: IOdspCache,
+        snapshotOptions: {[key: string]: number},
         isFirstTimeDocumentOpened = true,
     ): Promise<IDocumentService> {
         let odspResolvedUrl: IOdspResolvedUrl = resolvedUrl as IOdspResolvedUrl;
         const options = odspResolvedUrl.createNewOptions;
         if (options) {
-            const templogger: ITelemetryLogger = ChildLogger.create(logger, "OdspDriver");
-            const event = PerformanceEvent.start(templogger,
+            const event = PerformanceEvent.start(logger,
                 {
                     eventName: "CreateNew",
                     isWithSummaryUpload: false,
@@ -103,60 +100,9 @@ export class OdspDocumentService implements IDocumentService {
             deltasFetchWrapper,
             socketIOClientP,
             cache,
+            snapshotOptions,
             isFirstTimeDocumentOpened,
         );
-    }
-
-    public static async createContainer(
-        createNewSummary: ISummaryTree,
-        createNewResolvedUrl: IResolvedUrl,
-        logger: ITelemetryBaseLogger | undefined,
-        cache: IOdspCache,
-        getStorageToken: (siteUrl: string, refresh: boolean) => Promise<string | null>,
-        factory: IDocumentServiceFactory,
-        storageFetchWrapper: IFetchWrapper,
-    ): Promise<IDocumentService> {
-        ensureFluidResolvedUrl(createNewResolvedUrl);
-        let odspResolvedUrl = createNewResolvedUrl as IOdspResolvedUrl;
-        const [, queryString] = odspResolvedUrl.url.split("?");
-
-        const searchParams = new URLSearchParams(queryString);
-        const filePath = searchParams.get("path");
-        if (!filePath) {
-            throw new Error("File path should be provided!!");
-        }
-        const newFileParams: INewFileInfo = {
-            driveId: odspResolvedUrl.driveId,
-            siteUrl: odspResolvedUrl.siteUrl,
-            filePath,
-            filename: odspResolvedUrl.fileName,
-        };
-
-        const templogger: ITelemetryLogger = ChildLogger.create(logger, "OdspDriver");
-
-        const event = PerformanceEvent.start(templogger,
-            {
-                eventName: "CreateNew",
-                isWithSummaryUpload: true,
-            });
-        try {
-            odspResolvedUrl = await createNewFluidFile(
-                getStorageToken,
-                newFileParams,
-                cache,
-                storageFetchWrapper,
-                createNewSummary);
-            const props = {
-                docId: odspResolvedUrl.hashedDocumentId,
-            };
-
-            const docService = factory.createDocumentService(odspResolvedUrl, logger);
-            event.end(props);
-            return docService;
-        } catch (error) {
-            event.cancel(undefined, error);
-            throw error;
-        }
     }
 
     private storageManager?: OdspDocumentStorageManager;
@@ -186,17 +132,18 @@ export class OdspDocumentService implements IDocumentService {
         public readonly odspResolvedUrl: IOdspResolvedUrl,
         getStorageToken: (siteUrl: string, refresh: boolean) => Promise<string | null>,
         getWebsocketToken: (refresh) => Promise<string | null>,
-        logger: ITelemetryBaseLogger | undefined,
+        logger: ITelemetryLogger,
         private readonly storageFetchWrapper: IFetchWrapper,
         private readonly deltasFetchWrapper: IFetchWrapper,
         private readonly socketIOClientP: Promise<SocketIOClientStatic>,
         private readonly cache: IOdspCache,
+        private readonly snapshotOptions: {[key: string]: number},
         private readonly isFirstTimeDocumentOpened = true,
     ) {
         this.joinSessionKey = `${this.odspResolvedUrl.hashedDocumentId}/joinsession`;
         this.isOdc = isOdcOrigin(new URL(this.odspResolvedUrl.endpoints.snapshotStorageUrl).origin);
         this.logger = ChildLogger.create(logger,
-            "OdspDriver",
+            undefined,
             {
                 odc: this.isOdc,
             });
@@ -260,6 +207,7 @@ export class OdspDocumentService implements IDocumentService {
             true,
             this.cache,
             this.isFirstTimeDocumentOpened,
+            this.snapshotOptions,
         );
 
         return new OdspDocumentStorageService(this.storageManager);
